chore: TravelNewsEditorListItemView 분리
